* add codepage parameter to all exported conversion functions,   should be the iconv name of the Windows codepage, currently   only used in Notes/Description field * add VCALENDAR container to VEVENT and VTODO containers * fix some compiler warnings * add codepage option to all conversion test programs, fix   parameter handling and update man pages
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@3713 599ab11a-0111-0410-8a50-d7a5b0953570

+++ librra/src/rra-decode.c
@@ -188,7 +188,7 @@ bool decode_database_stream(uint8_t* buffer, unsigned id)
 		switch (propvals[i].propid & 0xff)
 		{
 			case CEVT_I2: printf("0x%08x  %i", propvals[i].val.iVal, propvals[i].val.iVal); break;
-			case CEVT_I4: printf("0x%08lx  %li", propvals[i].val.lVal, propvals[i].val.lVal); break;
+			case CEVT_I4: printf("0x%08lx  %li", (long unsigned)propvals[i].val.lVal, (long)propvals[i].val.lVal); break;
 
 			case CEVT_R8:  break;
 			case CEVT_BOOL:  break;
